Save image optimized otherwise save image without enhancement
Some large JPEG images (say 3000x3000) would trigger an IOError
exception when PIL engine is storing the image. The solution is to
increase ImageFile.MAXBLOCK. This modification allows to save the image
without enhancement when it is bigger than ImageFile.MAXBLOCK, so the
image is stored anyway, but it logs a message in order to hint one to
increase the value of ImageFile.MAXBLOCK for fine tuning.