Add ability for engines to do cleanup on shutdown
The use case for this is when an engine relies on something
that needs to be cleaned up on application shutdown.

Currently there is no way for an engine to be aware of
the application lifecycle. This seems like the most
lightweight way to let and engine know that application
shutdown is happening.