fixing _composite.apply with bigger source images
_composite.apply is pasting an image A into an image B at a virtual
position in space but not expanding the size of image B. If the image A
is bigger than image B and should be placed at x < 0 and/or y < 0 then
the image isn't using all the available space in image B in case image A
is bigger.

This is fixed now. The added test was failing before applying the
bugfix.