Pass through querystring params via new `HTTP_LOADER_FORWARD_QUERYSTRING*` settings.
This allows you to generate thumbnails for any number of private AWS S3
buckets with a single Thumbor server, as long as you pass absolute
signed AWS S3 URLs to Thumbor.

This follows the pattern established with the
`HTTP_LOADER_FORWARD_*HEADERS*` settings.

Case is ignored when comparing parameter names against the whitelist.

To prevent tampering, querystring parameters being passed through are
not allowed to override or extend existing querystring parameters in the
urlencoded original image URL.

NOTE:

Any querystring parameters that are not temporary/ephemeral, and which
affect the original image to be downloaded, should be specified as part
of the urlencoded original image URL, NOT as part of the Thumbor URL.

Querystring parameters in the Thumbor URL are (still) not considered to
be part of the original image URL, which is used as a cache key, which
would mean multiple requests for the same image with different
signatures would be downloaded and processed multiple times.

A consequence of this is that you should NOT allow "unsafe" Thumbor URLs
when passing through auth credentials as querystring parameters. Once
Thumbor has successfully processed and cached the image for such a
request, subsequent requests could access the cached image without
providing any auth credentials.

Instead, you are expected to rely on the HMAC validation for "safe"
Thumbor URLs, and ensure that clients are authorized before generating
safe URLs for private images that require auth credentials.

The same is true for auth credentials passed through as request headers.

Likewise, a similar issue exists when allowing unsafe URLs with the AWS
S3 loader, where Thumbor itself generates the AWS S3 signatures.