Removing undocumented distributed collage filter
The filter has been moved to
https://github.com/globocom/thumbor-distributed-collage-filter and is
already available at PyPI, so users can keep using it as soon as they
upgrade thumbor to this release.

This filter should not be in thumbor's codebase as it has dependencies
to numpy and opencv. If any portion of thumbor's codebase has optional
dependencies, they should be explicitly handled, and that's not the case
in this filter. Other than that, it was in the default set of filters.

Thumbor's committers, please strive to reduce thumbor's codebase, not
increase it. We worked hard to ensure that Thumbor is as extensible as
it can be. The reasoning here is that people can write thumbor
extensions and use them with thumbor's lightweight core. There's no reason to
increase thumbor's codebase. We've also done it in the past (I'm looking
at you both Redis and Mongo storages), but we've been trying hard to reduce
these dependencies, as they make evolving thumbor much harder.