[IMPR] Removed callbacks for async/await
This commit removes all callbacks from the codebase and modernizes
it to use async methods and await for their results.

As a side-effect of the modernizing I also tried to fix all lint
errors in the codebase. There are still some in there, but most
should be fixed by now.

The changes introduced in this commit are not backwards compatible.
All the storages, loaders, result storages, filters and detectors
need to be made async to work with thumbor's next major release.

The decision was hard, but as we turned to python3 several of the
extensions were not going to be compatible anyway, so it seemed
like a good timing to break compatibility and modernize the codebase.

When the release is finished, there'll be a release notes detailing
the migration that needs to be done in order to be migrate existing
extensions.