[FEAT] URL Routers support (#1261)
`Is Backwards Compatible`: yes

This commit adds support for a new class of plugins: Routers.
Routers are responsible for integrating new routes in thumbor's
App.

This is the missing extensiblity part in thumbor and one of the
reasons scope has crept in it (upload support).
After this has been released we'll start separating some of these
into their own modules.

One example apart from upload is healthcheck. While thumbor will
ship with a default healthcheck that's backwards compatible,
it will allow plugin builders to implement their own healthcheck
module (verifies if services that thumbor depends on are
available).

Creating a router is as simple as extending an abstract
`BaseRouter` and implementing the required methods.

Configuring which routers to use should be very familiar to any
thumbor user and is just a config value with the fully qualified
name of the class.