[FIX] Fixed hard dependency on pycurl
**Is backwards compatible**: yes

Thanks to @tony612 for pointing that thumbor 7.0.0a1 introduced
a hard dependency on pycurl (we import it in http_loader), but it
is described as a dev dependency in our pyproject.toml.

The correct behavior is that it should be an optional dependency
and this was a regression.

This commit fixes this behavior and hopefully puts the simple
async http client on par with the main functionality in pycurl
http client (most advanced usages still require pycurl).