[FEAT] thumbor-doctor
**Is backwards compatible**: yes

Thumbor doctor is a console app that users can run to verify their
thumbor install. That way they can send it with their issues in order to
make it easier for contributors to investigate what's going on.

The idea is that we keep adding checks to thumbor doctor over time.
It can be refactored to make it easier to add a new check, but this will
do for now.