[FIX] Make fill with blur engine-agnostic
The `fill` filter, when used with color set to "blur", depended on the
engine having the `blur` method. That broke the filter when the engine
in use doesn't fulfill such dependency.

By using the already existing blur capabilities of the `blur` filter,
`fill(blur)` now wokrs for engines other than PIL.