[IMPR] Compatibility loader,storage,res storage
**Is backwards compatible**: yes

With this change we introduce compatibility loader, storage and
result storage.

Users can now keep using the libraries they are already used to
(like tc_aws) to load, store and cache results.

At the same time, thumbor's codebase can keep clean with async
functions in all these structures.

We have, intentionally, skipped providing compatibility layers for
filters and detectors as these are not in the way of adopting
thumbor in Python 3, but may review this decision in the future.