ext/filters/_convolution: Apply convolution to alpha channel too
Not applying convolution to the alpha channel of images with
transparency results in an image where the colors are blurred but the
alpha mask is not. This change adds the alpha channel to the mix – when
it exists – fixinng the issue.

Both `blur` and `fill` filters are affected, hence the new tests and
fixtures.