Remove snapshottest to reduce number of dependencies (#1433)
Having an extra package that can be replaced with something already
included makes packaging easier. For instance, in Debian, one would have
to either be fortunate to find an existing package or go over the
trouble of creating such package and all its dependencies.

I believe this CL is a good small compromise considering the benefit it
brings.